+++ Source/Core/DolphinQML/Changes.md
@@ -5,3 +5,4 @@ This document serves to note the known changes compared to DolphinQt
 ## Header
 
 - If the window is too small to display all the toolbar icons, a scrollbar will apear instead of a chevron with an overflow menu
+- The about dialog doesn't use a native pop-up window but instead an in-app overlay